pyproject.toml
[tool.poetry]
name = "aiocqlengine"
version = "0.0.1"
description = ""
authors = ["Darren <charact3@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
aiocassandra = "^2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-asyncio = "^0.10.0"
pytest-cov = "^2.8"
pylava = "^0.2.2"
yapf = "^0.28.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
